refactor: move main merge logic from df to reshape package
